3-deployment/app/requirements.txt
fastapi==0.100.0
pydantic>=1.8.0,<2.0.0
uvicorn==0.22.0

pytest==6.2.5
requests==2.25.1
sqlmodel==0.0.16
asyncpg==0.28.0

joblib==1.4.2
pandas==2.2.2
xlrd
skforecast

requests
psycopg2-binary
python-dotenv

matplotlib